Move system-notification module and service to shared/notification/system-notification
