+++ src/main/webapp/app/admin/system-notification-management/system-notification-management.component.ts
@@ -13,7 +13,7 @@ import { onError } from 'app/shared/util/global.utils';
 import { Subject } from 'rxjs';
 import { SystemNotification } from 'app/entities/system-notification.model';
 import { ITEMS_PER_PAGE } from 'app/shared/constants/pagination.constants';
-import { SystemNotificationService } from 'app/core/system-notification/system-notification.service';
+import { SystemNotificationService } from 'app/shared/notification/system-notification/system-notification.service';
 import { AlertService } from 'app/core/alert/alert.service';
 
 @Component({
