+++ src/main/webapp/app/app.module.ts
@@ -3,7 +3,7 @@ import './vendor.ts';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ArtemisSystemNotificationModule } from 'app/core/system-notification/system-notification.module';
+import { ArtemisSystemNotificationModule } from 'app/shared/notification/system-notification/system-notification.module';
 import { NavbarComponent } from 'app/shared/layouts/navbar/navbar.component';
 import { NotificationSidebarComponent } from 'app/shared/notification/notification-sidebar/notification-sidebar.component';
 import { PageRibbonComponent } from 'app/shared/layouts/profiles/page-ribbon.component';
