+++ src/main/webapp/app/shared/notification/system-notification/system-notification.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { ArtemisSharedModule } from 'app/shared/shared.module';
-import { SystemNotificationService } from 'app/core/system-notification/system-notification.service';
+import { SystemNotificationService } from 'app/shared/notification/system-notification/system-notification.service';
 
 @NgModule({
     imports: [ArtemisSharedModule],
