+++ src/main/webapp/app/shared/notification/system-notification/system-notification.service.ts
